added sysio_get and filled in beta rule
changed current sysio function to sysio_set and added a new sysio_get
function that will return a dictionary of filenames from the output of
cmd for each extension that is specified. copied over beta rule code
and made modifications to get shared summary filename using the
sysio_get function. this can serve as an example for future use of the
sysio_get function